✨ Update due date format and styling in AssignmentList, AssignmentPageCard components
🔧 Adjust due date format in AssignmentList and update styling in AssignmentPageCard
components for better readability. Remove unused imports in SubmitedFiles
component and refactor to use SelectAssignmentCard instead of SelectAssignment.
Include assignment details in GET request for submitted files.
