+++ online-assignment/components/_component/SubmitedFiles.tsx
@@ -1,8 +1,7 @@
 "use client"
 
-import { useGetAssignmentsQuery } from "@/lib/features/roomSlice";
-import { Dispatch, SetStateAction, useState } from "react";
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "../ui/select";
+import { useState } from "react";
+import SelectAssignmentCard from "./SelectAssignmentCard";
 import SubmitedFilesCard from "./SubmitedFilesCard";
 
 type Props = {
@@ -14,7 +13,7 @@ const SubmitedFiles = ({ student_id }:Props) => {
 
     return (
     <div className="w-full md:container gap-3 mx-auto flex flex-col">
-        <SelectAssignment
+        <SelectAssignmentCard
             setAssignment_id={setAssignment_id}
             assignment_id={assignment_id}
             student_id={student_id}
@@ -25,37 +24,6 @@ const SubmitedFiles = ({ student_id }:Props) => {
         />
     </div>
   )
-}
-
-
-type SelectAssignmentProps = {
-    student_id:string;
-    assignment_id:string
-    setAssignment_id:Dispatch<SetStateAction<string>>;
-}
-
-const SelectAssignment = ({ student_id,assignment_id,setAssignment_id }:SelectAssignmentProps)=>{
-    const { data:assignments } = useGetAssignmentsQuery(student_id);
-    return(
-        <Select
-        onValueChange={(assignment_id)=>{
-            setAssignment_id(assignment_id)
-        }} 
-        defaultValue={assignment_id}
-    >
-        <SelectTrigger>
-            <SelectValue placeholder="Select assignment" />
-        </SelectTrigger>
-        <SelectContent>
-            {
-                assignments&&
-                assignments.map((assignment) =>(
-                    <SelectItem key={assignment?.id} value={assignment?.id!}>{assignment.title}</SelectItem>
-                ))
-            }
-        </SelectContent>
-    </Select> 
-    )
 }
