+++ online-assignment/types/types.d.ts
@@ -75,4 +75,5 @@ declare type SubmittedFile = {
     fileUrl: string;
     fileKey: string;
     submittedAt: Date;
+    assignment?:Assignment
 }
